Move reservation info modal to shared folder
