+++ app/assets/styles/app.less
@@ -34,7 +34,6 @@
 @import './loader';
 @import './panel';
 @import './react-select';
-@import './reservation-info-modal';
 
 // Customizations
 @import './customization/espoo/customization';
