+++ app/shared/modals/modals.less
@@ -1,1 +1,2 @@
 @import './comment/comment-modal.less';
+@import './reservation-info/reservation-info-modal';
