+++ app/shared/modals/reservation-info/index.js
@@ -1,0 +1,3 @@
+import ReservationInfoModalContainer from './ReservationInfoModalContainer';
+
+export default ReservationInfoModalContainer;
