+++ app/shared/modals/reservation-info/reservation-info-modal.less
